# `fs.Dir.deleteTree` optimizations and recursive implementation