@andrewrk 2022-10-05
> * Is a recursive implementation okay? An alternative with the same/similar performance could use allocation instead of recursion.

I'd like to fully exhaust the non-recursive, non-heap-allocating implementation options before resorting to this. Could you try a version that uses a fixed-size stack buffer (and resorts to redundant syscall operations in the case that the buffer is exhausted)?

> * Should the non-recursive/non-allocating version be maintained? If so, which should be the default? How should they be named?

I think we should have only one implementation in the standard library. Hopefully it can accomplish all the characteristics we want:
 * Equivalent or better perf of `rm -rf`.
 * No possibility of stack overflow due to "user input" of a deeply nested file system tree
 * No allocator parameter required.

If it's not possible to satisfy all these requirements then we can revisit the question, but I'm not convinced that will be necessary.