@squeek502 2022-10-05
That makes total sense @andrewrk, submitted that as https://github.com/ziglang/zig/pull/13073.